service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pgvector==0.3.3
sentence-transformers==3.1.1
openai==1.51.0
pyyaml==6.0.2
python-multipart==0.0.10
httpx==0.27.2
structlog==24.4.0
tenacity==9.0.0
numpy==1.26.4

# Sprint 1 (A1, A2, A4)
slowapi==0.1.9
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.20.0
pyjwt==2.9.0

# Dev / test
pytest==8.3.3
pytest-asyncio==0.24.0
